+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.1-alpha.171](https://github.com/DIG-Network/dig-chia-sdk/compare/v0.0.1-alpha.170...v0.0.1-alpha.171) (2024-10-26)
+
+
+### Features
+
+* add interval promise ([380f5da](https://github.com/DIG-Network/dig-chia-sdk/commit/380f5dae168256e54cc9f0618e6aaa753546b2e5))
+* optionally use redis cache instead of node cache if the right env is set up ([9c48999](https://github.com/DIG-Network/dig-chia-sdk/commit/9c4899987512a9ed4984dd6d31f327c0ed162237))
+
 ### [0.0.1-alpha.170](https://github.com/DIG-Network/dig-chia-sdk/compare/v0.0.1-alpha.169...v0.0.1-alpha.170) (2024-10-10)